deploy/installer: Refactor pkg naming
